+++ models/operation.go
@@ -17,24 +17,24 @@ type Operation struct {
 	Pkh                 string      `json:"pkh"`
 	Secret              string      `json:"secret"`
 	Source              string      `json:"source"`
-	Fee                 float64     `json:"fee"`
-	Counter             float64     `json:"counter"`
-	GasLimit            float64     `json:"gas_limit"`
-	StorageLimit        float64     `json:"storage_limit"`
+	Fee                 int64     `json:"fee"`
+	Counter             int64     `json:"counter"`
+	GasLimit            int64     `json:"gas_limit"`
+	StorageLimit        int64     `json:"storage_limit"`
 	PublicKey           string      `json:"public_key"`
-	Amount              float64     `json:"amount"`
+	Amount              int64     `json:"amount"`
 	Destination         string      `json:"destination"`
 	Parameters          string      `json:"parameters"`
 	ManagerPubkey       string      `json:"manager_pubkey"`
-	Balance             float64     `json:"balance"`
+	Balance             int64       `json:"balance"`
 	Spendable           bool        `json:"spendable"`
 	Delegatable         bool        `json:"delegatable"`
 	Script              string      `json:"script"`
 	Storage             string      `json:"storage"`
 	Status              string      `json:"status"`
-	ConsumedGas         float64     `json:"consumed_gas"`
-	StorageSize         float64     `json:"storage_size"`
-	PaidStorageSizeDiff float64     `json:"paid_storage_size_diff"`
+	ConsumedGas         int64     `json:"consumed_gas"`
+	StorageSize         int64     `json:"storage_size"`
+	PaidStorageSizeDiff int64     `json:"paid_storage_size_diff"`
 	OriginatedContracts string      `json:"originated_contracts"`
 	BlockHash           null.String `json:"block_hash"`
 	BlockLevel          null.Int    `json:"block_level"`
